pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["civitas"]

[project]
name = "civitas"
version = "0.1.0"
description = "The production runtime for Python agents"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "msgpack>=1.1",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13",
]

[project.urls]
Homepage = "https://github.com/jerynmathew/python-civitas"
Repository = "https://github.com/jerynmathew/python-civitas"
Documentation = "https://github.com/jerynmathew/python-civitas#readme"
"Bug Tracker" = "https://github.com/jerynmathew/python-civitas/issues"

[project.optional-dependencies]
otel      = [
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]
anthropic = ["anthropic>=0.20"]
openai    = ["openai>=1.0"]
gemini    = ["google-generativeai>=0.8"]
mistral   = ["mistralai>=1.0"]
litellm   = ["litellm>=1.40"]
zmq       = ["pyzmq>=25"]
nats      = ["nats-py>=2.6"]
fiddler   = ["fiddler-client>=3.0"]
dev       = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "anthropic>=0.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "types-pyyaml>=6.0.12",
]
docs      = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-gen-files>=0.5",
]

[project.scripts]
civitas = "civitas.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=civitas --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
branch = true
omit = [
    # Requires live infra (ZMQ IPC/TCP, NATS server) — covered in test_m2_1/m2_2
    "civitas/transport/zmq.py",
    "civitas/transport/nats.py",
    "civitas/transport/__init__.py",   # protocol definition only
    # Require live API keys — unit-tested via mocks
    "civitas/plugins/anthropic.py",
    "civitas/plugins/openai.py",
    "civitas/plugins/gemini.py",
    "civitas/plugins/mistral.py",
    "civitas/plugins/litellm.py",
    # Fiddler integration is a planned stub with no callable implementation yet
    "civitas/plugins/fiddler.py",
    # Adapter coverage lives in test_m2_6_adapters.py (excluded to avoid
    # inflating the threshold with tests that mock optional deps)
    "civitas/adapters/*",
    # CLI and dashboard require Typer/Rich setup — covered in test_m3_1/m3_3
    "civitas/cli/*",
    "civitas/dashboard/*",
    # Entry point shim
    "civitas/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@overload",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "ASYNC",
]
ignore = [
    "E501",
    # ASYNC109: timeout as a named parameter is intentional public API for ask()/request().
    # Callers pass it as a kwarg; using asyncio.timeout() internally is already done.
    "ASYNC109",
]

[tool.ruff.lint.isort]
known-first-party = ["civitas"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = false
disallow_untyped_defs = true
disallow_any_generics = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["msgpack.*", "zmq.*", "nats.*", "litellm.*", "openai.*", "google", "google.*", "mistralai.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["agents", "agents.*"]
ignore_missing_imports = true
